fix: 빌드오류 수정으로 suspense추가
